9.22 firstCommit

+++ app/src/main/java/com/example/coinpocket/domain/use_case/amount/UpdateAmount.kt
@@ -14,7 +14,7 @@ class UpdateAmount (
                                 day: String,
                                 categoryImage:CategoryImage,
                                 content:String,
-                                amount:Int,
+                                amount:Long,
                                 ){
         repository.updateAmount(
             id,
